Added Email functionality for the host

+++ src/main/java/com/wissen/service/impl/OTPServiceImpl.java
@@ -1,23 +1,34 @@
 package com.wissen.service.impl;
 
 import java.time.LocalDateTime;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Random;
 
-import com.wissen.exceptions.VisitorManagementException;
+import javax.transaction.Transactional;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
 import com.wissen.constants.Constants;
+import com.wissen.constants.enums.EmailTemplates;
+import com.wissen.dto.EmailDTO;
 import com.wissen.dto.OtpDTO;
+import com.wissen.entity.Employee;
 import com.wissen.entity.OTP;
 import com.wissen.entity.Timing;
 import com.wissen.entity.Visitor;
+import com.wissen.exceptions.VisitorManagementException;
 import com.wissen.repository.OTPRepository;
 import com.wissen.repository.TimingRepository;
 import com.wissen.repository.VisitorRepository;
 import com.wissen.service.OTPService;
-
-import javax.transaction.Transactional;
 
 /**
  * 
@@ -42,11 +53,27 @@ public class OTPServiceImpl implements OTPService {
 	@Autowired
 	private EmailService emailService;
 
+	@Value("${email.subject}")
+	private String otpEmailSubject;
+
+	@Value("${host.email.subject}")
+	private String hostEmailSubject;
+
+	/**
+	 * This method fetches the Visitor details based on the given input either email
+	 * id or phone number. It generates a 4 digit OTP and store it in visitor_otp
+	 * table and then send an email to the visitor with OTP
+	 * 
+	 * @param phEmail : input email id or phone number to fetch visitor details
+	 * 
+	 * @return : validation message
+	 * 
+	 */
 	@Override
 	@Transactional
 	public String sendOTP(String phEmail) {
-		String response = "";
 		Visitor visitor = visitorRepository.findByEmailOrPhoneNumber(phEmail);
+
 		if (visitor == null)
 			throw new VisitorManagementException(Constants.VISITOR_DOESNOT_EXISTS);
 
@@ -57,14 +84,19 @@ public class OTPServiceImpl implements OTPService {
 		if (Objects.isNull(visitorOTPRecord)) {
 			String otp = generateOTP();
 			storeOTP(visitor, otp);
-			emailService.sendEmail(visitor.getFullName(), visitor.getEmail(), otp);
-		} else{
-			emailService.sendEmail(visitor.getFullName(), visitor.getEmail(), visitorOTPRecord.getOtp());
-		}
-		response = Constants.OTP_GENERATION_MESSAGE;
-		return response;
+			prepareAndSendEmailForOTP(visitor, otp);
+		} else
+			prepareAndSendEmailForOTP(visitor, visitorOTPRecord.getOtp());
+		return Constants.OTP_GENERATION_MESSAGE;
 	}
 
+	/**
+	 * This method stores an OTP in visitor_otp table with expiry time as current
+	 * time + 15 mins
+	 * 
+	 * @param visitor
+	 * @param otp
+	 */
 	private void storeOTP(Visitor visitor, String otp) {
 		OTP visitorOtp = new OTP();
 		visitorOtp.setVisitor(visitor);
@@ -73,22 +105,64 @@ public class OTPServiceImpl implements OTPService {
 		otpRepository.save(visitorOtp);
 	}
 
+	/**
+	 * This method prepares to send an email to visitor's email id with generated
+	 * OTP
+	 * 
+	 * @param visitor : Visitor details
+	 * @param otp     : generated OTP
+	 */
+	private void prepareAndSendEmailForOTP(Visitor visitor, String otp) {
+		try {
+			EmailDTO emailDTO = new EmailDTO();
+
+			Map<String, Object> model = new HashMap<String, Object>();
+			model.put(Constants.VISITOR_NAME, visitor.getFullName());
+			model.put(Constants.VISITOR_EMAIL, visitor.getEmail());
+			model.put(Constants.OTP, otp);
+			emailDTO.setProperties(model);
+
+			List<String> recipients = new ArrayList<>();
+			recipients.add(visitor.getEmail());
+			emailDTO.setRecipients(recipients);
+			emailDTO.setSubject(otpEmailSubject);
+			emailDTO.setTemplateName(EmailTemplates.VISITOR_OTP.getTemplateName());
+
+			emailService.sendEmail(emailDTO);
+		} catch (Exception e) {
+			throw new VisitorManagementException(
+					"Exception Occurred while preparing the temple to send an email for OTP");
+		}
+
+	}
+
+	/**
+	 * This method validates the generated OTP and returns the success message else
+	 * throw an exception if validation fails
+	 * 
+	 * @param data : DTO object contains OTP and phone/email
+	 * 
+	 * @return : validation message
+	 */
 	@Override
 	@Transactional
 	public String verifyOTP(OtpDTO data) {
-		String response = "";
 		OTP visitorOTPRecord = otpRepository.findByVisitorEmailOrPhoneNumber(data.getPhEmail());
 		if (visitorOTPRecord != null) {
 			if (data.getOtp().equalsIgnoreCase(visitorOTPRecord.getOtp())) {
-				response = Constants.VALID_OTP_MESSAGE;
-				setTimings(visitorOTPRecord);
-				otpRepository.delete(visitorOTPRecord);
+
+				Visitor visitor = visitorRepository.findByEmailOrPhoneNumber(data.getPhEmail());
+				if (visitor != null) {
+					setTimings(visitor);
+					otpRepository.delete(visitorOTPRecord);
+					return Constants.VALID_OTP_MESSAGE + ". " + sendEmailToHost(visitor);
+				} else
+					throw new VisitorManagementException(Constants.VISITOR_NOT_FOUND_MESSAGE);
 			} else
 				throw new VisitorManagementException(Constants.INVALID_OTP_MESSAGE);
 		} else
 			throw new VisitorManagementException(Constants.NEW_OTP_MESSAGE);
 
-		return response;
 	}
 
 	/**
@@ -107,14 +181,12 @@ public class OTPServiceImpl implements OTPService {
 	 * This method sets the timings for the visitor in timing table with inTime as
 	 * current time and outTime as null
 	 * 
-	 * @param otpRecord : to fetch the visitor details based on visitor Id in OTP table
+	 * @param visitor : Visitor record
 	 */
-	private void setTimings(OTP otpRecord) {
+	private void setTimings(Visitor visitor) {
 
-		Optional<Visitor> visitorOpt = visitorRepository.findByVisitorId(otpRecord.getVisitor().getVisitorId());
+		if (visitor != null) {
 
-		if (visitorOpt.isPresent()) {
-			Visitor visitor = visitorOpt.get();
 			// Sort the timings list based on the outTime in descending order
 			Collections.sort(visitor.getTimings(), Comparator.comparing(Timing::getOutTime).reversed());
 
@@ -126,6 +198,56 @@ public class OTPServiceImpl implements OTPService {
 			timing.setVisitorType(visitor.getTimings().get(0).getVisitorType());
 			timingRepository.save(timing);
 
+		}
+	}
+
+	/**
+	 * This method sort the visitor details based on the out time and get the latest
+	 * entry details to fetch the employee details based on the employee id and
+	 * retrieve the host details to send an intimation email for visitor
+	 * 
+	 * @param visitor : input visitor details
+	 * @return : validation message
+	 */
+	private String sendEmailToHost(Visitor visitor) {
+		Collections.sort(visitor.getTimings(), Comparator.comparing(Timing::getOutTime).reversed());
+		Employee employee = visitor.getTimings().get(0).getEmployee();
+		if (!Objects.isNull(employee)) {
+			prepareAndSendEmailForHost(visitor, employee);
+			return Constants.HOST_EMAIL_SENT_MESSAGE;
+		} else
+			throw new VisitorManagementException(Constants.HOST_NOT_FOUND_MESSAGE);
+	}
+
+	/**
+	 * This method fetches the employee details from employee table based on the
+	 * employee id and retrieve the host details to send an intimation email for
+	 * visitor
+	 * 
+	 * @param visitor  : input visitor details
+	 * @param employee : Employee object to fetch host details
+	 * @return : validation message
+	 */
+	private void prepareAndSendEmailForHost(Visitor visitor, Employee employee) {
+		try {
+			EmailDTO emailDTO = new EmailDTO();
+
+			Map<String, Object> model = new HashMap<String, Object>();
+			model.put(Constants.VISITOR_NAME, visitor.getFullName());
+			model.put(Constants.EMPLOYEE_NAME, employee.getFirstName());
+
+			emailDTO.setProperties(model);
+
+			List<String> recipients = new ArrayList<>();
+			recipients.add(employee.getEmail());
+			emailDTO.setRecipients(recipients);
+			emailDTO.setSubject(hostEmailSubject);
+			emailDTO.setTemplateName(EmailTemplates.HOST_INTIMATION.getTemplateName());
+
+			emailService.sendEmail(emailDTO);
+		} catch (Exception e) {
+			throw new VisitorManagementException(
+					"Exception Occurred while preparing the temple to send an email for Host");
 		}
 	}
